Updating test folder to include test_parse_data.py
