+++ fitbenchmarking/logs/fitbenchmarking.log
@@ -1,20 +1,6 @@
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/ENGINX193749_calibration_peak19.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/ENGINX193749_calibration_peak20.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/ENGINX193749_calibration_peak23.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/ENGINX193749_calibration_peak5.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/ENGINX193749_calibration_peak6.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/ENGINX236516_vanadium_bank1_10brk.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/ENGINX236516_vanadium_bank1_20brk.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/EVS14188-90_Gaussian_peaks_1.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/EVS14188-90_Gaussian_peaks_2.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/GEMpeak1.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak1.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak2.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak3.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak4.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak5.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak6.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak7.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak8.txt
-[13:53:29]  INFO: /home/mwathen/fitbenchmarking/benchmark_problems/Neutron_data/WISH17701_peak9.txt
-[13:53:29]  INFO: * Testing fitting of problem ENGINX 193749 calibration, spectrum 651, peak 19
+[14:00:36]  ERROR: Could not find equation, index went out of bounds!
+[14:00:36]  INFO: *** Loading dat data file Misra1a.dat ***
+[14:00:36]  INFO: 17 lines were ignored in this problem file.
+If any problems occur, please uncomment line above this print to display the full output.
+[14:00:36]  INFO: 17 lines were ignored in this problem file.
+If any problems occur, please uncomment line above this print to display the full output.
